app/requirements.txt
aiobotocore==2.23.0
boto3==1.38.27
botocore==1.38.27
# catboost==1.2.2
fastapi==0.111.0
implicit==0.7.2
# jupyterlab
# lightfm==1.17
# mlflow==2.7.1
# optuna==3.4.0
pandas==2.1.1
prometheus-fastapi-instrumentator==7.1.0
psycopg==3.1.12
psycopg[binary,pool]
psycopg2-binary==2.9.10
pyarrow==13.0.0
python-dotenv==1.1.1
requests==2.31.0
s3fs==2025.5.1
scikit-learn==1.3.2
# scikit-surprise==1.1.4
scipy==1.11.3
# seaborn==0.13.0
uvicorn==0.24.0.post1
